// src/types/resources.ts
export type OrdersPageResourcesProps = {
  currentOrders: string;
  pastOrders: string;
  orderNow: string;
  noOrdersYet: string;
  orderNo: string;
  viewDetails: string;
  reorder: string;
  trackOrder: string;
  items: string;
  item: string;
  orderType: string;
};